Fix missed warnings when compiling
AIInterface.cpp(3528): warning C4018: '<': signed/unsigned mismatch)

+++ src/world/Units/Creatures/AIInterface.cpp
@@ -3525,7 +3525,7 @@ void AIInterface::UpdateAISpells()
         // Shuffle Around our Spells to randomize the cast
         if (mCreatureAISpells.size())
         {
-            for (int i = 0; i < mCreatureAISpells.size() - 1; ++i)
+            for (size_t i = 0; i < mCreatureAISpells.size() - 1; ++i)
             {
                 int j = i + rand() % (mCreatureAISpells.size() - i);
                 std::swap(mCreatureAISpells[i], mCreatureAISpells[j]);
@@ -3848,7 +3848,7 @@ void AIInterface::sendStoredText(definedEmoteVector store, Unit* target)
     // Shuffle Around our textIds to randomize it
     if (store.size())
     {
-        for (int i = 0; i < store.size() - 1; ++i)
+        for (size_t i = 0; i < store.size() - 1; ++i)
         {
             int j = i + rand() % (store.size() - i);
             std::swap(store[i], store[j]);
